requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
amplitude-analytics==1.1.5
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
decorator==5.2.1
distro==1.9.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.8.2
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.0.2
openai==1.65.1
pillow==10.4.0
proglog==0.1.10
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
redis==5.2.1
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
yarl==1.18.3
